Refactor lib.align
