--- docs/full/lib/aligner.rst
@@ -1,7 +1,0 @@
-aligner module
-==============
-
-.. automodule:: lib.aligner
-   :members:
-   :undoc-members:
-   :show-inheritance:
