--- docs/full/lib/alignments.rst
@@ -1,7 +1,0 @@
-alignments module
-=================
-
-.. automodule:: lib.alignments
-   :members:
-   :undoc-members:
-   :show-inheritance:
