--- docs/full/lib/faces_detect.rst
@@ -1,21 +1,0 @@
-********************
-faces\_detect module
-********************
-
-Handles detected and aligned faces objects and their associated masks.
-
-.. rubric:: Module Summary
-
-.. autosummary::
-   :nosignatures:
-   
-   ~lib.faces_detect.BlurMask
-   ~lib.faces_detect.DetectedFace
-   ~lib.faces_detect.Mask
-
-.. rubric:: Module
-
-.. automodule:: lib.faces_detect
-   :members:
-   :undoc-members:
-   :show-inheritance:
