+++ plugins/extract/mask/_base.py
@@ -82,7 +82,7 @@ class Masker(Extractor):  # pylint:disable=abstract-method
         to ``dict`` for internal processing.
 
         To ensure consistent batch sizes for masker the items are split into separate items for
-        each :class:`~lib.faces_detect.DetectedFace` object.
+        each :class:`~lib.align.DetectedFace` object.
 
         Remember to put ``'EOF'`` to the out queue after processing
         the final batch
@@ -91,7 +91,7 @@ class Masker(Extractor):  # pylint:disable=abstract-method
         :attr:`~plugins.extract._base.Extractor.batchsize`:
 
         >>> {'filename': [<filenames of source frames>],
-        >>>  'detected_faces': [[<lib.faces_detect.DetectedFace objects]]}
+        >>>  'detected_faces': [[<lib.align.DetectedFace objects]]}
 
         Parameters
         ----------
