+++ tools/manual/frameviewer/editor/mask.py
@@ -176,7 +176,7 @@ class Mask(Editor):
 
         Parameters
         ----------
-        mask: :class:`lib.faces_detect.Mask`
+        mask: :class:`lib.align.Mask`
             The mask object
         mask_scale: float
             The scaling factor from the stored mask size to the internal mask size
@@ -333,7 +333,7 @@ class Mask(Editor):
     def _update_roi_box(self, mask, face_index, color):
         """ Update the region of interest box for the current mask.
 
-        mask: :class:`~lib.faces_detect.Mask`
+        mask: :class:`~lib.align.Mask`
             The current mask object to create an ROI box for
         face_index: int
             The index of the face within the current frame
