+++ tools/preview/preview.py
@@ -15,13 +15,13 @@ import cv2
 import numpy as np
 from PIL import Image, ImageTk
 
-from lib.aligner import Extract as AlignerExtract
+from lib.align import Extract as AlignerExtract
 from lib.cli.args import ConvertArgs
 from lib.gui.utils import get_images, get_config, initialize_config, initialize_images
 from lib.gui.custom_widgets import Tooltip
 from lib.gui.control_helper import ControlPanel, ControlPanelOption
 from lib.convert import Converter
-from lib.faces_detect import DetectedFace
+from lib.align import DetectedFace
 from lib.multithreading import MultiThread
 from lib.utils import FaceswapError
 from lib.queue_manager import queue_manager
@@ -215,7 +215,7 @@ class Samples():
 
     @property
     def alignments(self):
-        """ :class:`~lib.alignments.Alignments`: The alignments for the preview faces """
+        """ :class:`~lib.align.Alignments`: The alignments for the preview faces """
         return self._alignments
 
     @property
